refactor: extract error boundary and rename existing one
