+++ packages/slice-machine/src/features/navigation/Navigation.tsx
@@ -3,7 +3,7 @@ import { useRouter } from "next/router";
 import { Suspense } from "react";
 
 import { telemetry } from "@/apiClient";
-import { ErrorBoundary } from "@/ErrorBoundary";
+import { DefaultErrorBoundary } from "@/errorBoundaries";
 import { CUSTOM_TYPES_CONFIG } from "@/features/customTypes/customTypesConfig";
 import { CUSTOM_TYPES_MESSAGES } from "@/features/customTypes/customTypesMessages";
 import { RepositoryInfo } from "@/features/navigation/RepositoryInfo";
@@ -79,20 +79,20 @@ export function Navigation() {
           />
         </ActionList>
 
-        <ErrorBoundary>
+        <DefaultErrorBoundary>
           <Suspense>
             <UpdateInfo />
           </Suspense>
-        </ErrorBoundary>
+        </DefaultErrorBoundary>
       </Box>
 
       <Box flexDirection="column">
         <ActionList variant="compact">
-          <ErrorBoundary>
+          <DefaultErrorBoundary>
             <Suspense>
               <OnboardingGuide />
             </Suspense>
-          </ErrorBoundary>
+          </DefaultErrorBoundary>
           <NavigationItem
             title="Documentation"
             href={documentationLink}
@@ -112,11 +112,11 @@ export function Navigation() {
             Icon={LightningIcon}
             active={router.asPath.startsWith("/changelog")}
             RightElement={
-              <ErrorBoundary>
+              <DefaultErrorBoundary>
                 <Suspense fallback={<Skeleton height={16} />}>
                   <SliceMachineVersion />
                 </Suspense>
-              </ErrorBoundary>
+              </DefaultErrorBoundary>
             }
           />
         </ActionList>
